refactor: split tableComponentNode file into smaller modules
